Use int64 for issueNumber parameter
Per review feedback: both issue_id and issue_number are integer in the
OpenAPI spec, so issueNumber should be int64 for consistency.

+++ github/issues_dependencies.go
@@ -20,7 +20,7 @@ type IssueDependencyRequest struct {
 // GitHub API docs: https://docs.github.com/rest/issues/issue-dependencies#list-dependencies-an-issue-is-blocked-by
 //
 //meta:operation GET /repos/{owner}/{repo}/issues/{issue_number}/dependencies/blocked_by
-func (s *IssuesService) ListBlockedBy(ctx context.Context, owner, repo string, issueNumber int, opts *ListOptions) ([]*Issue, *Response, error) {
+func (s *IssuesService) ListBlockedBy(ctx context.Context, owner, repo string, issueNumber int64, opts *ListOptions) ([]*Issue, *Response, error) {
 	u := fmt.Sprintf("repos/%v/%v/issues/%v/dependencies/blocked_by", owner, repo, issueNumber)
 	u, err := addOptions(u, opts)
 	if err != nil {
@@ -46,7 +46,7 @@ func (s *IssuesService) ListBlockedBy(ctx context.Context, owner, repo string, i
 // GitHub API docs: https://docs.github.com/rest/issues/issue-dependencies#add-a-dependency-an-issue-is-blocked-by
 //
 //meta:operation POST /repos/{owner}/{repo}/issues/{issue_number}/dependencies/blocked_by
-func (s *IssuesService) AddBlockedBy(ctx context.Context, owner, repo string, issueNumber int, issueDepReq IssueDependencyRequest) (*Issue, *Response, error) {
+func (s *IssuesService) AddBlockedBy(ctx context.Context, owner, repo string, issueNumber int64, issueDepReq IssueDependencyRequest) (*Issue, *Response, error) {
 	u := fmt.Sprintf("repos/%v/%v/issues/%v/dependencies/blocked_by", owner, repo, issueNumber)
 	req, err := s.client.NewRequest("POST", u, issueDepReq)
 	if err != nil {
@@ -67,7 +67,7 @@ func (s *IssuesService) AddBlockedBy(ctx context.Context, owner, repo string, is
 // GitHub API docs: https://docs.github.com/rest/issues/issue-dependencies#remove-dependency-an-issue-is-blocked-by
 //
 //meta:operation DELETE /repos/{owner}/{repo}/issues/{issue_number}/dependencies/blocked_by/{issue_id}
-func (s *IssuesService) RemoveBlockedBy(ctx context.Context, owner, repo string, issueNumber int, issueID int64) (*Issue, *Response, error) {
+func (s *IssuesService) RemoveBlockedBy(ctx context.Context, owner, repo string, issueNumber int64, issueID int64) (*Issue, *Response, error) {
 	u := fmt.Sprintf("repos/%v/%v/issues/%v/dependencies/blocked_by/%v", owner, repo, issueNumber, issueID)
 	req, err := s.client.NewRequest("DELETE", u, nil)
 	if err != nil {
@@ -88,7 +88,7 @@ func (s *IssuesService) RemoveBlockedBy(ctx context.Context, owner, repo string,
 // GitHub API docs: https://docs.github.com/rest/issues/issue-dependencies#list-dependencies-an-issue-is-blocking
 //
 //meta:operation GET /repos/{owner}/{repo}/issues/{issue_number}/dependencies/blocking
-func (s *IssuesService) ListBlocking(ctx context.Context, owner, repo string, issueNumber int, opts *ListOptions) ([]*Issue, *Response, error) {
+func (s *IssuesService) ListBlocking(ctx context.Context, owner, repo string, issueNumber int64, opts *ListOptions) ([]*Issue, *Response, error) {
 	u := fmt.Sprintf("repos/%v/%v/issues/%v/dependencies/blocking", owner, repo, issueNumber)
 	u, err := addOptions(u, opts)
 	if err != nil {
